requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
conan==1.60.0
defusedxml==0.7.1
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-mail==1.4.1
fasteners==0.18
greenlet==3.0.3
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jira==3.5.1
MarkupSafe==2.1.3
node-semver==0.6.1
oauthlib==3.2.2
openstep-parser==1.5.4
orjson==3.9.7
packaging==23.1
paho-mqtt==1.6.1
patch-ng==1.17.4
pbxproj==3.5.0
pluginbase==1.0.1
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.15.1
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==5.3.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
